move `AdditionalAttributes` and `TextProp` into core
